Move api into its own module
