+++ src/jetstream/api/apikeys.go
@@ -1,4 +1,4 @@
-package interfaces
+package api
 
 import "time"
