+++ src/jetstream/api/general.go
@@ -1,4 +1,4 @@
-package interfaces
+package api
 
 import "github.com/labstack/echo/v4"
