+++ src/jetstream/api/go.mod
@@ -1,0 +1,29 @@
+module github.com/cloudfoundry-incubator/stratos/src/jetstream/api
+
+go 1.21.0
+
+require (
+	github.com/cloudfoundry-incubator/stratos/src/jetstream/repository/interfaces v0.0.0-20191015083202-4c24d26ae1ec
+	github.com/gorilla/sessions v1.2.1
+	github.com/gorilla/websocket v1.5.0
+	github.com/govau/cf-common v0.0.7
+	github.com/labstack/echo/v4 v4.11.1
+	github.com/sirupsen/logrus v1.9.3
+)
+
+require (
+	github.com/cloudfoundry-community/go-cfenv v1.17.0 // indirect
+	github.com/gorilla/securecookie v1.1.1 // indirect
+	github.com/labstack/gommon v0.4.0 // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.19 // indirect
+	github.com/mitchellh/mapstructure v1.1.2 // indirect
+	github.com/onsi/ginkgo v1.16.5 // indirect
+	github.com/onsi/gomega v1.27.10 // indirect
+	github.com/valyala/bytebufferpool v1.0.0 // indirect
+	github.com/valyala/fasttemplate v1.2.2 // indirect
+	golang.org/x/crypto v0.11.0 // indirect
+	golang.org/x/net v0.12.0 // indirect
+	golang.org/x/sys v0.10.0 // indirect
+	golang.org/x/text v0.11.0 // indirect
+)
