+++ src/jetstream/api/portal_proxy.go
@@ -1,4 +1,4 @@
-package interfaces
+package api
 
 import (
 	"database/sql"
