+++ src/jetstream/api/tokens.go
@@ -1,4 +1,4 @@
-package interfaces
+package api
 
 // Token -
 type Token struct {
