+++ src/jetstream/api/websocket.go
@@ -1,4 +1,4 @@
-package interfaces
+package api
 
 import (
 	"fmt"
